requirements.txt
duckdb>=0.9.0
ollama>=0.1.0
pydantic>=2.5.0
pyyaml>=6.0.1
rich>=13.7.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.0
